Fixes #240 - Add I18n support for admin interface
This adds the possibility of having admin iterface
translated. The language is chosen based
on the language of the user that is currently
logged in. If no user is logged in,
the default is English.

Note that this does not interfere with the actual
language of the site when the site ApplicationController
sets the locale from params as usual.

I added the translations for Czech.